% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map_to_Genes.R
\name{Evidence_Symbols}
\alias{Evidence_Symbols}
\title{Prepares adds gene symbols to evidence file object}
\usage{
Evidence_Symbols(evi = NULL, Species = "Human")
}
\arguments{
\item{evi}{Evidence table created with \code{\link[proteus]{readEvidenceFile}}.}

\item{Species}{indicating the species. "Human" (default),
"Mouse", or "Yeast" are the only options.}
}
\value{
data.frame with uniprot ids replaced with gene "SYMBOL".
For yeast, "GENENAME" is returned.
}
\description{
Prepares adds gene symbols to evidence file object
}
\examples{
library(proteusLabelFree)
evi<-proteusLabelFree::evi

evi$protein<-gsub(".*[sp|]([^.]+)[|].*", "\\\\1", evi$protein)
evi_symbols<-ProtGene::Evidence_Symbols(evi=evi, Species="Yeast")
head(evi_symbols)
}
