% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_Evidence.R
\name{Proteus_Prep}
\alias{Proteus_Prep}
\title{Prepares metadata file object for Proteus}
\usage{
Proteus_Prep(evi = NULL, measure.cols = "Intensity")
}
\arguments{
\item{evi}{Evidence table created with \code{\link[proteus]{readEvidenceFile}}.}

\item{measure.cols}{A named list of measure columns; should be the same as
used in \code{\link[proteus]{readEvidenceFile}}}
}
\value{
data.frame with sample metadata
}
\description{
Prepares metadata file object for Proteus
}
\examples{
devtools::install_github("bartongroup/proteusLabelFree")
library(proteusLabelFree)
evidenceFile <- system.file("extdata", "evidence.txt.gz", package="proteusLabelFree")
evi <- readEvidenceFile(evidenceFile)
Meta<-Proteus_Prep(evi=evi, measure.cols="intensity")
Meta
}
