% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Making_ExpressionSet.R
\name{eset_proteus}
\alias{eset_proteus}
\title{Converts proteins table and metadata into eset_proteus}
\usage{
eset_proteus(Proteins_table = NULL, MS_metadata = NULL)
}
\arguments{
\item{Proteins_table}{Output of makeProteinTable function}

\item{MS_metadata}{Metadata object prepared by Proteus_Prep function}
}
\value{
Expressset
}
\description{
Converts proteins table and metadata into eset_proteus
}
\examples{
library(proteusLabelFree)
data(proteusLabelFree)

evi$protein<-gsub(".*[sp|]([^.]+)[|].*", "\\\\1", evi$protein)
evi_symbols<-ProtGene::Evidence_Symbols(evi=evi, Species="Yeast")
head(evi_symbols)

Meta<-Proteus_Prep(evi=evi_symbols, measure.cols="intensity")
Meta$condition<-gsub("-[1-9]","", Meta$experiment)
Meta$measure<-"Intensity"

pepdat <- proteus::makePeptideTable(evi_symbols, Meta, ncores=4)
Proteinsdat <- proteus::makeProteinTable(pepdat, ncores = 4)

data_se<-eset_proteus(Proteins_table=Proteinsdat, MS_metadata=Meta)

DEP::plot_numbers(data_se)
}
