% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map_to_Genes.R
\name{Evidence_Symbols}
\alias{Evidence_Symbols}
\title{Prepares adds gene symbols to evidence file object}
\usage{
Evidence_Symbols(evi = NULL, Species = "Human")
}
\arguments{
\item{evi}{Evidence table created with \code{\link[proteus]{readEvidenceFile}}.}

\item{Species}{indicating the species. "Human" (default),
"Mouse", or "Yeast" are the only options.}
}
\value{
data.frame with uniprot ids replaced with gene "SYMBOL".
For yeast, "GENENAME" is returned.
}
\description{
Prepares adds gene symbols to evidence file object
}
\examples{
# Getting evidence file
library(ProtGene)
library(proteusLabelFree)
evidenceFile <- system.file("extdata", "evidence.txt.gz", package="proteusLabelFree")

# The directory of the evidence file is required
Evidence_input <- proteus::readEvidenceFile(evidenceFile)

# extracting uniprot ids: "sp|P00175|CYB2_YEAST" to "P00175
# Your data may not need this step
Evidence_input$protein<-gsub(".*[sp|]([^.]+)[|].*", "\\\\1", Evidence_input$protein)
head(Evidence_input)

# Converting to gene name
evi_symbols<-ProtGene::Evidence_Symbols(evi=Evidence_input, Species="Yeast")
head(evi_symbols)
}
