% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaData_Builder.R
\name{Proteus_Prep}
\alias{Proteus_Prep}
\title{Prepares metadata file object for Proteus}
\usage{
Proteus_Prep(evi = NULL, measure.cols = "Intensity")
}
\arguments{
\item{evi}{Evidence table created with \code{\link[proteus]{readEvidenceFile}}.}

\item{measure.cols}{A named list of measure columns; should be the same as
used in \code{\link[proteus]{readEvidenceFile}}}
}
\value{
data.frame with sample metadata
}
\description{
Prepares metadata file object for Proteus
}
\examples{
# Getting evidence file
library(ProtGene)
library(proteusLabelFree)
evidenceFile <- system.file("extdata", "evidence.txt.gz", package="proteusLabelFree")

# The directory of the evidence file is required
Evidence_input <- proteus::readEvidenceFile(evidenceFile)

# extracting uniprot ids: "sp|P00175|CYB2_YEAST" to "P00175
# Your data may not need this step
Evidence_input$protein<-gsub(".*[sp|]([^.]+)[|].*", "\\\\1", Evidence_input$protein)
head(Evidence_input)

# Converting to gene name
evi_symbols<-ProtGene::Evidence_Symbols(evi=Evidence_input, Species="Yeast")
head(evi_symbols)

# Generating metadata from evidence file
Meta<-Proteus_Prep(evi=evi_symbols)
head(Meta)
}
